Moved settings to own package
